Refactor Bluesky types into modular structure
- Split Bluesky-related types into separate files under `src/bluesky/types/`
- Create index files for easier module exports
- Reorganize imports in `bluesky.ts` to use new type structure
- Add index files for other modules to improve import consistency
