+++ src/bluesky/types/EmbeddedMedia.ts
@@ -1,0 +1,4 @@
+import { VideoEmbed } from "./VideoEmbed";
+import { ImagesEmbed } from "./ImagesEmbed";
+
+export type EmbeddedMedia = VideoEmbed | ImagesEmbed;
