+++ src/bluesky/types/PostRecord.ts
@@ -1,0 +1,16 @@
+import { AppBskyFeedPost, AppBskyRichtextFacet } from "@atproto/api";
+import { EmbeddedMedia } from "./EmbeddedMedia";
+
+export interface PostRecord extends Partial<AppBskyFeedPost.Record> {}
+
+export class PostRecordImpl implements PostRecord {
+  readonly $type = "app.bsky.feed.post";
+  [k: string]: unknown;
+
+  constructor(
+    public text: string,
+    public createdAt: string,
+    public facets: AppBskyRichtextFacet.Main[],
+    public embed: EmbeddedMedia
+  ) {}
+}
